Introduce alembic
Alembic should help us do database migrations easily while deploying new
Artemis code.

This initial version is autogenerated for the current database version.

This is the initial point, from now on we will generate database
migrations so we can nicely upgrade our staging/production env database
which has persistent storage.

+++ server/nominishift-develop.sh
@@ -2,6 +2,11 @@
 
 set -x
 
+if [ "$(readlink -f $(dirname 0))" != "$PWD" ]; then
+    echo "This script must be run from project root"
+    exit 1
+fi
+
 trap 'kill $(jobs -p)' EXIT
 
 export ARTEMIS_CONFIG_DIR="$(pwd)/configuration"
@@ -32,6 +37,7 @@ if [ "$ARTEMIS_WORKER_QUEUES" != "" ]; then
     ARTEMIS_WORKER_OPTIONS="-Q \"${ARTEMIS_WORKER_QUEUES}\" ${ARTEMIS_WORKER_OPTIONS}"
 fi
 
+poetry run alembic upgrade head
 poetry run artemis-init-postgres-schema
 
 poetry run artemis-api-server &
